LoLBets/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
aiohttp>=3.9.1
Pillow>=10.2.0
beautifulsoup4>=4.12.0
cloudscraper>=1.2.71
Brotli>=1.1.0
